create new sub directory called cloud
