+++ wi/wi-extension/src/choreo-rpc/connection.ts
@@ -19,7 +19,7 @@
 import { type ChildProcessWithoutNullStreams, spawn } from "child_process";
 import { type MessageConnection, StreamMessageReader, StreamMessageWriter, createMessageConnection } from "vscode-jsonrpc/node";
 import { ext } from "../extensionVariables";
-import { parseJwt } from "../utils";
+import { parseJwt } from "../utils/commonUtils";
 import { getChoreoExecPath } from "./cli-install";
 
 export class StdioConnection {
