+++ wi/wi-extension/src/cloud/cmds/cmd-utils.ts
@@ -18,7 +18,7 @@
 
 import { WICommandIds, type ComponentKind, type ExtensionName, type Organization, type Project, type UserInfo } from "@wso2/wso2-platform-core";
 import { ProgressLocation, type QuickPickItem, QuickPickItemKind, type WorkspaceFolder, commands, window, workspace } from "vscode";
-import { type ExtensionVariables, ext } from "../extensionVariables";
+import { type ExtensionVariables, ext } from "../../extensionVariables";
 import { dataCacheStore } from "../stores/data-cache-store";
 import { webviewStateStore } from "../stores/webview-state-store";
