+++ wi/wi-extension/src/cloud/cmds/refresh-directory-context-cmd.ts
@@ -18,7 +18,7 @@
 
 import { WICommandIds, type ICmdParamsBase } from "@wso2/wso2-platform-core";
 import { type ExtensionContext, commands, window } from "vscode";
-import { ext } from "../extensionVariables";
+import { ext } from '../../extensionVariables';
 import { contextStore } from "../stores/context-store";
 import { isRpcActive, setExtensionName } from "./cmd-utils";
