+++ wi/wi-extension/src/cloud/git/encoding.ts
