+++ wi/wi-extension/src/cloud/git/github/credentialProvider.ts
