+++ wi/wi-extension/src/cloud/git/ssh-askpass-empty.sh
