+++ wi/wi-extension/src/cloud/git/ssh-askpass.sh
