+++ wi/wi-extension/src/cloud/git/terminal.ts
