+++ wi/wi-extension/src/cloud/stores/context-store.ts
@@ -35,9 +35,9 @@ import * as yaml from "js-yaml";
 import { ProgressLocation, window, workspace } from "vscode";
 import { createStore } from "zustand";
 import { persist } from "zustand/middleware";
-import { ext } from "../extensionVariables";
+import { ext } from '../../extensionVariables';
 import { getGitRemotes, getGitRoot } from "../git/util";
-import { isSamePath, isSubpath } from "../utils";
+import { isSamePath, isSubpath } from '../../utils';
 import { dataCacheStore } from "./data-cache-store";
 import { locationStore } from "./location-store";
 import { getWorkspaceStateStore } from "./store-utils";
