+++ wi/wi-extension/src/cloud/stores/location-store.ts
