+++ wi/wi-extension/src/extensionVariables.ts
@@ -19,8 +19,8 @@
 import type { WSO2Terminologies } from "@wso2/wso2-platform-core";
 import * as vscode from "vscode";
 import type { ChoreoRPCClient } from "./choreo-rpc";
-import { defaultTerminologies, webviewStateStore } from "./stores/webview-state-store";
-import type { WSO2AuthenticationProvider } from "./auth/wso2-auth-provider";
+import { defaultTerminologies, webviewStateStore } from "./cloud/stores/webview-state-store";
+import type { WSO2AuthenticationProvider } from "./cloud/auth/wso2-auth-provider";
 
 /**
  * Extension context wrapper
