+++ wi/wi-extension/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./commonUtils";
+export * from "./pathUtils";
